fix: don't crash when changing the language (only when offline)
This change only makes it work when in offline/airplane mode.
When connected to a server it still crashes.
Likely due to a different part of the code.

Thanks to https://github.com/piplup55 for pointing out to me that this worked in old versions.
I then tested a few older builds and found out that it started to not work with version `v1.1.0`.
Using `git bisect` I was able to find out that commit 34691affb0d0fa3e42f39e524d946b5b83b782b9 introduced the issue.